Move address features to crate::stream
Puts them in a similar place to DuplexAddr.
